discount percentages stored inside integers
Changed all discount percentages to be stored inside integers because they never get generated as floats

+++ include/Objects/Discounts.hpp
@@ -8,14 +8,14 @@ class Discount
 {
 protected:
     size_t discountID; // unique
-    float discountPercent;
+    int discountPercent;
 
 public:
-    Discount(size_t id, float percent);
+    Discount(size_t id, int percent);
 
     virtual DiscountType getType() const;
     virtual void view() const;
-    float getPercent();
+    int getPercent();
     size_t getID() const;
 
     virtual ~Discount() = default;
@@ -26,7 +26,7 @@ class BonusDiscount : public Discount
     float bonus;
 
 public:
-    BonusDiscount(size_t id, float percent, float bonus);
+    BonusDiscount(size_t id, int percent, float bonus);
 
     DiscountType getType() const override;
     void view() const override;
@@ -38,7 +38,7 @@ class BrandDiscount : public Discount
     std::string brandName;
 
 public:
-    BrandDiscount(size_t id, float percent, const std::string &brandName);
+    BrandDiscount(size_t id, int percent, const std::string &brandName);
 
     DiscountType getType() const override;
     void view() const override;
